requirements.txt
joblib==1.2.0
numpy==1.23.4
pandas==1.5.0
python-dateutil==2.8.2
pytz==2022.4
scikit-learn==1.1.2
scipy==1.9.2
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
python-dotenv
pydantic
bcrypt==4.1.2
sqlalchemy
pydantic_settings
passlib==1.7.4
pyjwt==2.10.1
fastapi==0.121.0
uvicorn==0.38.0